refactor: refactor of code in the app
