+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:nested/nested.dart';
 import 'package:provider/provider.dart';
 
 import 'package:epicticker/config/theme/app_theme.dart';
-import 'package:epicticker/presentation/provider/count_down_provider.dart';
+import 'package:epicticker/presentation/providers/crud/count_down_provider.dart';
 import 'package:epicticker/presentation/routes/main_routes.dart';
 
 void main() => runApp(const MyApp());
@@ -16,7 +16,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MultiProvider(
       providers: <SingleChildWidget>[
-        ChangeNotifierProvider<CountDownProvider>(create: (BuildContext context) => CountDownProvider())
+        ChangeNotifierProvider<CountDownProvider>(create: (_) => CountDownProvider())
       ],
       child: MaterialApp(
         title: 'EpicTicker',
