+++ lib/presentation/providers/countdown/count_down_state.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class CountDownState with ChangeNotifier {
+	late TextEditingController textNameController = TextEditingController();
+	late TextEditingController fullDateController = TextEditingController();
+
+	void disposeController() {
+		textNameController.dispose();
+		textNameController.dispose();
+	}
+}
