+++ lib/presentation/widgets/dashboard_widget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:epicticker/common/color.dart';
 import 'package:epicticker/common/text_styles.dart';
-import 'package:epicticker/presentation/provider/count_down_provider.dart';
+import 'package:epicticker/presentation/providers/crud/count_down_provider.dart';
 import 'package:epicticker/utils/get_month_name.dart';
 import 'package:epicticker/utils/get_week_name.dart';
